First attempt at question typing
We need question types to define what answers look like -- I'm thinking this needs to be done in terms of underlying variables. We can insist later that the request for Booleans is actually served up as a request for a String, defining a Boolean output, that can be parsed by math.js.